+++ src/Unitarum/Options.php
@@ -10,6 +10,8 @@ class Options implements OptionsInterface
 {
     private $fixtureFolder;
     private $dsn;
+    private $dbUserName;
+    private $dbPassword;
 
     public function __construct(array $options)
     {
@@ -18,6 +20,12 @@ class Options implements OptionsInterface
         }
         if (isset($options[self::DSN_OPTION])) {
             $this->setDsn($options[self::DSN_OPTION]);
+        }
+        if (isset($options[self::DB_USERNAME])) {
+            $this->setDbUserName($options[self::DB_USERNAME]);
+        }
+        if (isset($options[self::DB_PASSWORD])) {
+            $this->setDbPassword($options[self::DB_PASSWORD]);
         }
     }
 
@@ -54,6 +62,42 @@ class Options implements OptionsInterface
     public function setDsn($dsn): self
     {
         $this->dsn = $dsn;
+        return $this;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getDbUserName()
+    {
+        return $this->dbUserName;
+    }
+
+    /**
+     * @param mixed $dbUserName
+     * @return Options
+     */
+    public function setDbUserName($dbUserName)
+    {
+        $this->dbUserName = $dbUserName;
+        return $this;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getDbPassword()
+    {
+        return $this->dbPassword;
+    }
+
+    /**
+     * @param mixed $dbPassword
+     * @return Options
+     */
+    public function setDbPassword($dbPassword)
+    {
+        $this->dbPassword = $dbPassword;
         return $this;
     }
 }