+++ test/Unitarum/Adapter/SqliteAdapterTest.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace UnitarumTest\Adapter;
+
+
+class SqliteAdapterTest
+{
+
+}