+++ test/Unitarum/OptionsTest.php
@@ -12,13 +12,20 @@ class OptionsTest extends TestCase
     {
         $fixtureFolder = 'data';
         $dsn = 'dsn:';
+        $dbUserName = 'username';
+        $dbPassword = 'password';
+
         $params = [
             OptionsInterface::FIXTURE_FOLDER_OPTION => $fixtureFolder,
-            OptionsInterface::DSN_OPTION => $dsn
+            OptionsInterface::DSN_OPTION => $dsn,
+            OptionsInterface::DB_USERNAME => $dbUserName,
+            OptionsInterface::DB_PASSWORD => $dbPassword,
         ];
 
         $options = new Options($params);
         $this->assertEquals($fixtureFolder, $options->getFixtureFolder());
         $this->assertEquals($dsn, $options->getDsn());
+        $this->assertEquals($dbUserName, $options->getDbUserName());
+        $this->assertEquals($dbPassword, $options->getDbPassword());
     }
 }